requirements.txt
numpy 
nltk
matplotlib.pyplot
tqdm
collections
math
random
sklearn